requirements.txt
Django~=4.2.1
djangorestframework~=3.14.0
factory-boy~=3.2.1
